corrected a bug to reduce the number of moves needed for 5 elements

+++ library/libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: tmongell <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 17:33:02 by tmongell          #+#    #+#             */
-/*   Updated: 2022/05/02 16:24:56 by tmongell         ###   ########.fr       */
+/*   Updated: 2022/05/10 18:26:04 by tmongell         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long int	ft_atoi(const char *str)
 {
-	int	i;
-	int	nb;
-	int	sign;
+	long	int	i;
+	long	int	nb;
+	long	int	sign;
 
 	nb = 0;
 	sign = 1;
